Implement Listassoc and Nonassoc! (#47)
* feat: implement nonassoc
* feat: implement listassoc

In this PR, we add 2 helpers, binop.nonassoc and binop.listassoc. These are a little
bit tricky: The idea is that we use lexical precedence to force tree-sitter to continue
the parse rather than try to restart the same expression.

Example:
1 > 2 > 3

We make tree-sitter parse the 2 > as different tokens, one of which is only available in
a chain. We make the one in the chain have higher precedence, thus making it impossible to
re-enter the same rule, b/c it will never be able to see the > token which starts the
expression.

Naming + API are up for grabs, but IT WORKS

it makes generating the parser take much longer, though